requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
black==20.8b1
click==7.1.2
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.8
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2020.5
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
